This checkin does batching of TimelineAggregator chunk creates and invalidates/deletes.
collector/src/main/java/com/ning/arecibo/collector/guice/CollectorConfig.java:

    Add int parameter maxChunkIdsToInvalidateOrDelete, defaulting to
    1000.  This means that by default, we'll batch up chunk inserts
    and deletes or invalidates whenever the count of chunks aggregated
    exceeds 1000.

collector/src/main/java/com/ning/arecibo/collector/persistent/TimelineAggregator.java:

    Add some counters to distinguish between queued write operations
    and the actual writes.

    Add a pair of lists, one to hold new chunks produced by
    aggregation, and the other to hold the chunkIds of aggregated
    chunks to be deleted or marked not valid.  In
    aggregateHostSampleChunks(), add the new chunk and the ids of the
    chunks aggregated to these lists, and if the list of ids exceeds
    the threshold, call performWrites().

    When we're done with an aggregation level, of there are any writes
    queued, call performWrites().

    Add new method performWrites(), which in a single transaction,
    inserts the new chunks and deletes or marks not valid the old
    chunks.

    Fix bug: Remove redundant timelineChunksCreated.inc().

collector/src/main/java/com/ning/arecibo/collector/persistent/TimelineAggregatorDAO.java:

    In makeTimelineChunkValid(), change the chunkId to be a long.

util/src/main/resources/com/ning/arecibo/util/timeline/persistent/TimelineDAOQueries.sql.stg:

    In insertTimelineChunk, explicitly initialize aggregation_level
    and not_valid, rather than letting the table default be used.

+++ collector/src/main/java/com/ning/arecibo/collector/persistent/TimelineAggregatorDAO.java
@@ -44,7 +44,7 @@ public interface TimelineAggregatorDAO extends Transactional<TimelineAggregatorD
     int getLastInsertedId();
 
     @SqlUpdate
-    void makeTimelineChunkValid(@Bind("chunkId") final int chunkId);
+    void makeTimelineChunkValid(@Bind("chunkId") final long chunkId);
 
     @SqlUpdate
     void makeTimelineChunksInvalid(@BindIn("chunkIds") final List<Long> chunkIds);
